requirements.txt
alabaster==0.7.12
asgiref==3.5.2
Babel==2.10.3
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.19
flake8==5.0.4
idna==3.4
imagesize==1.4.1
importlib-metadata==5.0.0
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
parso==0.8.3
pluggy==1.0.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.5.0
pytz==2022.6
requests==2.28.1
snowballstemmer==2.2.0
sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
ujson==5.5.0
urllib3==1.26.12
zipp==3.10.0
